Display title of song first.
Apparently, this format also affects the listing in the browser
view. So, whenever the window size is too small (e.g. when I'm
using tmux), I cannot always see the name of the song that I am
about to select. Fix this by just swapping the arrangement of
the album and title.